switched to float

+++ tavily/async_tavily.py
@@ -279,7 +279,7 @@ class AsyncTavilyClient:
                include_images: bool = None,
                extract_depth: Literal["basic", "advanced"] = None,
                format: Literal["markdown", "text"] = None,
-               timeout: int = 150,
+               timeout: float = 150,
                include_favicon: bool = None,
                **kwargs
                ) -> dict:
@@ -351,7 +351,7 @@ class AsyncTavilyClient:
                     extract_depth: Literal["basic", "advanced"] = None,
                     include_images: bool = None,
                     format: Literal["markdown", "text"] = None,
-                    timeout: int = 150,
+                    timeout: float = 150,
                     include_favicon: bool = None,
                     **kwargs
                     ) -> dict:
@@ -391,7 +391,7 @@ class AsyncTavilyClient:
                exclude_domains: Sequence[str] = None,
                allow_external: bool = None,
                include_images: bool = None,
-               timeout: int = 150,
+               timeout: float = 150,
                **kwargs
                ) -> dict:
         """
@@ -457,7 +457,7 @@ class AsyncTavilyClient:
                     exclude_domains: Sequence[str] = None,
                     allow_external: bool = None,
                     include_images: bool = None,
-                    timeout: int = 150,
+                    timeout: float = 150,
                     **kwargs
                     ) -> dict:
         """
